split up style-guide into separate components
